+++ exercises/src/01 CompositeUI/after/Divergent.Finance.API/Controllers/PricesController.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.Web.Http;
 using Divergent.Finance.Data.Repositories;
+using System.Collections.Generic;
 
 namespace Finance.API.Controllers
 {
@@ -29,17 +30,19 @@ namespace Finance.API.Controllers
                         => prices.Single(s => s.ProductId == productId).ItemPrice);
         }
 
-        //[HttpGet, Route("orders/total/{orderIds}")]
-        //public async Task<dynamic> GetOrdersTotal(string orderIds)
-        //{
-        //    var _ids = productIds.Split(",".ToCharArray(), StringSplitOptions.RemoveEmptyEntries)
-        //           .Select(id => Guid.Parse(id))
-        //           .ToList();
+        [HttpGet, Route("orders/total")]
+        public async Task<IDictionary<Guid, double>> GetOrdersTotal(string orderIds)
+        {
+            var _ids = orderIds.Split(",".ToCharArray(), StringSplitOptions.RemoveEmptyEntries)
+                   .Select(id => Guid.Parse(id))
+                   .ToList();
 
-        //    var prices = await _financeRepository.Prices();
-
-        //    return _ids.Sum(productId
-        //                => prices.Single(s => s.ProductId == productId).ItemPrice);
-        //}
+            var repo = new OrdersTotalRepository();
+            var prices = (await repo.OrderTotalPrices())
+                .Where(otp=>_ids.Contains(otp.OrderId))
+                .ToDictionary(otp=>otp.OrderId, otp=>otp.TotalPrice);
+            
+            return prices;
+        }
     }
 }
