+++ exercises/src/01 CompositeUI/after/Divergent.Finance.Data/Models/OrderTotalPrice.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Divergent.Finance.Data.Models
+{
+    public class OrderTotalPrice
+    {
+        public Guid OrderId { get; set; }
+        public double TotalPrice { get; set; }
+    }
+}
