+++ exercises/src/01 CompositeUI/after/Divergent.Finance.Data/Repositories/OrdersTotalRepository.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Divergent.Finance.Data.Models;
+
+namespace Divergent.Finance.Data.Repositories
+{
+    public class OrdersTotalRepository
+    {
+        public async Task<List<OrderTotalPrice>> OrderTotalPrices()
+        {
+            return SeedOrderTotalPrices();
+        }
+
+        private List<OrderTotalPrice> SeedOrderTotalPrices()
+        {
+            var particularOrderId = Guid.Parse("6c3945c9-0b64-414a-9e5f-892de482ae2b");
+            var nserivceBusOrderId = Guid.Parse("41f74caf-2ef9-46ca-9445-2fcdf35751fb");
+
+            return new List<OrderTotalPrice>()
+            {
+                new OrderTotalPrice() { OrderId = particularOrderId, TotalPrice = 120 },
+                new OrderTotalPrice() { OrderId= nserivceBusOrderId, TotalPrice = 90 }
+            };
+        }
+    }
+}
