+++ exercises/src/01 CompositeUI/after/Divergent.Frontend/app/branding/orders/ordersView.html
@@ -8,16 +8,9 @@
                 Id #{{order.orderNumber}}
                 <br />
                 <i>Items count:</i> {{order.orderItemsCount}}
+                <br />
+                <strong>Order total:</strong> {{order.orderTotalPrice}}
             </div>
-            ﻿
-            <div>
-                <strong>Order total:</strong> {{order.totalPrice}}
-            </div>
-
-            <!--<div>
-                <strong>Customer</strong>
-                <div>{{order.customerName}} (#{{order.customerId}})</div>
-            </div>-->
         </li>
     </ul>
