+++ exercises/src/01 CompositeUI/after/Divergent.Frontend/index.html
@@ -68,9 +68,11 @@
 
     <!-- Finance module -->
     <script src="/app/modules/finance/_module.js" type="text/javascript"></script>
+    <script src="/app/modules/finance/ordersLoadedSubscriber.js" type="text/javascript"></script>
 
     <!-- Sales module -->
     <script src="/app/modules/sales/_module.js" type="text/javascript"></script>
+    <script src="/app/modules/sales/orderListAppender.js" type="text/javascript"></script>
     <script src="/app/modules/sales/orders/ordersController.js" type="text/javascript"></script>
 
 </body>
